+++ cherilibs/trunk/tools/build_sdk.sh
@@ -29,9 +29,11 @@
 # @BERI_LICENSE_HEADER_END@
 #
 
+SDK_VERSION=1
+
 check_dep()
 {
-	if [ x"`which $1`" == x ] ; then
+	if [ -z "`which $1`" ] ; then
 		echo error: No $1 binary found in PATH: ${PATH}
 		echo $2
 		FOUNDDEPS=0
@@ -53,10 +55,11 @@ check_dep cmake "Required for building LLVM"
 check_dep ninja "Required for building LLVM"
 check_dep clang++ "Required for building LLVM"
 check_dep git "Required for fetching source code"
-if [ x"${MAKEOBJDIRPREFIX}" == x ] ; then
+WD=`realpath .`
+if [ -z "${MAKEOBJDIRPREFIX}" ] ; then
 	echo "MAKEOBJDIRPREFIX is not set, use current location? (lots of object code will go there!)"
 	read yn
-	if [ x"$yn" == xy ] ; then
+	if [ "$yn" == y ] ; then
 		mkdir obj
 		export MAKEOBJDIRPREFIX=${WD}/obj
 	else
@@ -64,7 +67,7 @@ if [ x"${MAKEOBJDIRPREFIX}" == x ] ; then
 		FOUNDDEPS=0
 	fi
 fi
-if [ x"${JFLAG}" == x ] ; then
+if [ -z "${JFLAG}" ] ; then
 	JFLAG=`sysctl kern.smp.cpus | awk '{ printf "-j" $2 }'`
 	echo No JFLAG specified, defaulting to ${JFLAG}
 fi
@@ -75,9 +78,25 @@ echo All dependencies satisfied
 if [ ! -d sdk ] ; then
 	mkdir sdk
 fi
-WD=`realpath .`
-SYSROOT_DIR=${WD}/sdk/
+if [ ! -d sdk/sysroot ] ; then
+	mkdir sdk/sysroot
+fi
+SDKROOT_DIR=${WD}/sdk/
+SYSROOT_DIR=${SDKROOT_DIR}sysroot/
 CPUTYPE=mips
+
+if [ -f $SDKROOT_DIR/version ] ; then
+	EXISTING_SDK_VERSION=`cat $SDKROOT_DIR/version`
+else
+	EXISTING_SDK_VERSION=0
+fi
+
+# If we're building a new version of the SDK, delete the old one
+if [ $EXISTING_SDK_VERSION -ne $SDK_VERSION ] ; then
+	echo Deleting old SDK if one exists...
+	rm -rf $SDKROOT_DIR
+	mkdir -p $SYSROOT_DIR
+fi
 
 #
 # Choice of hard- vs soft-float is cached in .hardfloat.
@@ -92,7 +111,7 @@ else
 	echo $yn > .hardfloat
 fi
 
-if [ x"$yn" == xy ] ; then
+if [ "$yn" == y ] ; then
 	echo "Will build SDK for hard-float MIPS"
 	CPUTYPE=mipsfpu
 else
@@ -133,13 +152,21 @@ else
 	try_to_run git clone https://github.com/CTSRD-CHERI/clang
 fi
 cd ..
+
+# If we've got an older version of the SDK, then delete the LLVM build dir and
+# reconfigure it for the new location.
+if [ $EXISTING_SDK_VERSION -eq 0 ] ; then
+	echo Removing old LLVM build directory...
+	rm -rf Build
+fi
+
 if [ -d Build ] ; then
 	cd Build
 else
 	mkdir Build
 	cd Build
 	echo Configuring LLVM Build...
-	try_to_run cmake .. -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_C_COMPILER=clang -DCMAKE_BUILD_TYPE=Release -DDEFAULT_SYSROOT=${SYSROOT_DIR} -DLLVM_DEFAULT_TARGET_TRIPLE=cheri-unknown-freebsd -DCMAKE_INSTALL_PREFIX=${SYSROOT_DIR} -G Ninja
+	try_to_run cmake .. -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_C_COMPILER=clang -DCMAKE_BUILD_TYPE=Release -DDEFAULT_SYSROOT=${SYSROOT_DIR} -DLLVM_DEFAULT_TARGET_TRIPLE=cheri-unknown-freebsd -DCMAKE_INSTALL_PREFIX=${SDKROOT_DIR} -G Ninja
 fi
 echo Building LLVM...
 try_to_run ninja
@@ -147,10 +174,10 @@ echo Installing LLVM...
 try_to_run ninja install
 cd ../..
 # delete some things that we don't need...
-rm -rf ${SYSROOT_DIR}/lib/lib*
-rm -rf ${SYSROOT_DIR}/share
-rm -rf ${SYSROOT_DIR}/include
-rm -f ${SYSROOT_DIR}/lib/clang/3.*/include/std*
+rm -rf ${SDKROOT_DIR}/lib/lib*
+rm -rf ${SDKROOT_DIR}/share
+rm -rf ${SDKROOT_DIR}/include
+rm -f ${SDKROOT_DIR}/lib/clang/3.*/include/std*
 CHERIBSD_ROOT=`realpath .`/cheribsd
 if [ -d cheribsd ] ; then
 	echo Updating CheriBSD...
@@ -174,37 +201,47 @@ echo Building the toolchain...
 # don't change.
 CHERIROOT_OBJ="${MAKEOBJDIRPREFIX}/mips.mips64`realpath .`/tmproot"
 CHERITOOLS_OBJ="${MAKEOBJDIRPREFIX}/mips.mips64`realpath .`/tmp/usr/bin/"
-FBSD_BUILD_ARGS="TARGET=mips TARGET_ARCH=mips64 CPUTYPE=${CPUTYPE} CHERI_CC=${SYSROOT_DIR}/bin/clang -DDB_FROM_SRC -DNO_ROOT"
+CHERILIBEXEC_OBJ="${MAKEOBJDIRPREFIX}/mips.mips64`realpath .`/tmp/usr/libexec/"
+FBSD_BUILD_ARGS="-DCHERI CPUTYPE=${CPUTYPE} CHERI_CC=${SDKROOT_DIR}/bin/clang -DDB_FROM_SRC -DNO_ROOT -DNO_WERROR"
 echo Building FreeBSD base distribution...
-echo make ${JFLAG} ${FBSD_BUILD_ARGS} buildworld
-try_to_run make ${JFLAG} ${FBSD_BUILD_ARGS} buildworld
-echo Installing FreeBSD base distribution to ${CHERIROOT_OBJ}...
-rm -rf "${CHERIROOT_OBJ}"
-mkdir -p "${CHERIROOT_OBJ}"
-try_to_run make ${JFLAG} ${FBSD_BUILD_ARGS} DESTDIR="${CHERIROOT_OBJ}" installworld
+echo NO_BUILDWORLD value: ${NO_BUILDWORLD}
+if [ -z "${NO_BUILDWORLD}" ]; then
+	echo make ${JFLAG} ${FBSD_BUILD_ARGS} buildworld
+	# Do a non-parallel cleandir to work around build system bugs.
+	try_to_run make ${FBSD_BUILD_ARGS} cleandir
+	try_to_run make ${JFLAG} ${FBSD_BUILD_ARGS} buildworld
+	echo Installing FreeBSD base distribution to ${CHERIROOT_OBJ}...
+	rm -rf "${CHERIROOT_OBJ}"
+	mkdir -p "${CHERIROOT_OBJ}"
+	try_to_run make ${JFLAG} ${FBSD_BUILD_ARGS} DESTDIR="${CHERIROOT_OBJ}" installworld
+fi
 echo Populating SDK...
 cd ${SYSROOT_DIR}
-(cd "${CHERIROOT_OBJ}" && tar cf - --include="./lib/" --include="./usr/include/" --include="./usr/lib/" @METALOG) | tar xf -
+(cd "${CHERIROOT_OBJ}" && tar cf - --include="./lib/" --include="./usr/include/" --include="./usr/lib/" --include="./usr/libdata/" @METALOG) | tar xf -
 if [ $? -ne 0 ] ; then
 	exit 1
 fi
 echo Installing tools...
+mkdir -p ${SDKROOT_DIR}/bin
 TOOLS="as lint objdump strings addr2line c++filt crunchide gcc gcov nm readelf strip ld objcopy size"
 for TOOL in ${TOOLS} ; do
-	cp -f ${CHERITOOLS_OBJ}/${TOOL} ${SYSROOT_DIR}/bin/${TOOL}
+	cp -f ${CHERITOOLS_OBJ}/${TOOL} ${SDKROOT_DIR}/bin/${TOOL}
 done
-#if [ -x /usr/local/bin/mips64-freebsd-ld ]; then
-#	echo Replacing ld with cross-binutils version...
-#	ln -sf /usr/local/bin/mips64-freebsd-ld ${SYSROOT_DIR}/bin/ld
-#fi
+# GCC wants the cc1 and cc1plus tools to be in the directory specified by -B.
+# We must make this the same directory that contains ld for linking and
+# compiling to both work...
+for TOOL in cc1 cc1plus ; do
+	cp -f ${CHERILIBEXEC_OBJ}/${TOOL} ${SDKROOT_DIR}/bin/${TOOL}
+done
+cd ${SDKROOT_DIR}/bin
 TOOLS="${TOOLS} clang clang++ llvm-mc llvm-objdump llvm-readobj llvm-size llc"
 for TOOL in ${TOOLS} ; do
-	ln -fs $TOOL ${SYSROOT_DIR}/bin/cheri-unknown-freebsd-${TOOL}
-	ln -fs $TOOL ${SYSROOT_DIR}/bin/mips4-unknown-freebsd-${TOOL}
-	ln -fs $TOOL ${SYSROOT_DIR}/bin/mips64-unknown-freebsd-${TOOL}
+	ln -fs $TOOL cheri-unknown-freebsd-${TOOL}
+	ln -fs $TOOL mips4-unknown-freebsd-${TOOL}
+	ln -fs $TOOL mips64-unknown-freebsd-${TOOL}
 done
 echo Fixing absolute paths in symbolic links inside lib directory...
-echo | cat | cc -x c - -o ${SYSROOT_DIR}/bin/fixlinks <<EOF
+echo | cat | cc -x c - -o ${SDKROOT_DIR}/bin/fixlinks <<EOF
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <dirent.h>
@@ -264,9 +301,9 @@ int main(int argc, char **argv)
 }
 EOF
 cd ${SYSROOT_DIR}/usr/lib
-try_to_run ../../bin/fixlinks 
+try_to_run ${SDKROOT_DIR}/bin/fixlinks 
 echo Compiling cheridis helper...
-echo | cat | cc -DLLVM_PATH=\"${SYSROOT_DIR}/bin/\" -x c - -o ${SYSROOT_DIR}/bin/cheridis <<EOF
+echo | cat | cc -DLLVM_PATH=\"${SDKROOT_DIR}/bin/\" -x c - -o ${SDKROOT_DIR}/bin/cheridis <<EOF
 #include <stdio.h>
 #include <string.h>
 
@@ -293,6 +330,7 @@ int main(int argc, char** argv)
 	pclose(dis);
 }
 EOF
-echo Done.  Use ${SYSROOT_DIR}/bin/clang to compile code.
-echo Add --sysroot=${SYSROOT_DIR} -B${SYSROOT_DIR} to your CFLAGS
+echo $SDK_VERSION > $SDKROOT_DIR/version
+echo Done.  Use ${SDKROOT_DIR}/bin/clang to compile code.
+echo Add --sysroot=${SYSROOT_DIR} -B${SDKROOT_DIR}bin to your CFLAGS
 rm -f error.log
